requirements.txt
playwright>=1.44.0
openpyxl>=3.1.2
pyyaml>=6.0.1
deep-translator>=1.11.4
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
